LayoutTests/fast/css/overscroll-behavior-invalidate-if-disabled-expected.txt

Test overscrollBehavior should be invalidated if overscrollBehaviorEnabled is disabled

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'overscrollBehavior' in document.documentElement.style is false
PASS 'overscroll-behavior' in getComputedStyle(document.documentElement) is false
PASS successfullyParsed is true

TEST COMPLETE
